+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
 - Added Travis CI
+- Re-added rule34 command.